requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.16
asyncpg==0.27.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.22
python-dotenv==1.0.0
celery==5.3.1
redis>=4.6.0
aiohttp==3.8.5
python-whois==0.7.3
pydantic==1.10.13
email-validator==2.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
